Move byte store choose_provider to remote_provider
