+++ src/rust/engine/fs/store/Cargo.toml
@@ -28,9 +28,7 @@ madvise = { workspace = true }
 parking_lot = { workspace = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
-remote_provider_opendal = { path = "../../remote_provider/remote_provider_opendal" }
-remote_provider_reapi = { path = "../../remote_provider/remote_provider_reapi" }
-remote_provider_traits = { path = "../../remote_provider/remote_provider_traits" }
+remote_provider = { path = "../../remote_provider" }
 serde = { workspace = true }
 serde_derive = { workspace = true }
 sharded_lmdb = { path = "../../sharded_lmdb" }
