+++ src/rust/engine/fs/store/src/lib.rs
@@ -89,7 +89,7 @@ mod remote_tests;
 
 // Consumers of this crate shouldn't need to worry about the exact crate structure that comes
 // together to make a store.
-pub use remote_provider_traits::RemoteOptions;
+pub use remote_provider::RemoteOptions;
 
 pub struct LocalOptions {
   pub files_max_size_bytes: usize,
