+++ src/rust/engine/fs/store/src/remote.rs
@@ -9,45 +9,11 @@ use bytes::Bytes;
 use futures::Future;
 use hashing::Digest;
 use log::Level;
-use remote_provider_traits::{ByteStoreProvider, LoadDestination, RemoteOptions};
+use remote_provider::{
+  choose_byte_store_provider, ByteStoreProvider, LoadDestination, RemoteOptions,
+};
 use tokio::fs::File;
 use workunit_store::{in_workunit, ObservationMetric};
-
-// TODO: this is probably better positioned somewhere else
-pub const REAPI_ADDRESS_SCHEMAS: [&str; 4] = ["grpc://", "grpcs://", "http://", "https://"];
-
-async fn choose_provider(options: RemoteOptions) -> Result<Arc<dyn ByteStoreProvider>, String> {
-  let address = options.cas_address.clone();
-  if REAPI_ADDRESS_SCHEMAS.iter().any(|s| address.starts_with(s)) {
-    Ok(Arc::new(
-      remote_provider_reapi::byte_store::Provider::new(options).await?,
-    ))
-  } else if let Some(path) = address.strip_prefix("file://") {
-    // It's a bit weird to support local "file://" for a 'remote' store... but this is handy for
-    // testing.
-    Ok(Arc::new(remote_provider_opendal::Provider::fs(
-      path,
-      "byte-store".to_owned(),
-      options,
-    )?))
-  } else if let Some(url) = address.strip_prefix("github-actions-cache+") {
-    // This is relying on python validating that it was set as `github-actions-cache+https://...` so
-    // incorrect values could easily slip through here and cause downstream confusion. We're
-    // intending to change the approach (https://github.com/pantsbuild/pants/issues/19902) so this
-    // is tolerable for now.
-    Ok(Arc::new(
-      remote_provider_opendal::Provider::github_actions_cache(
-        url,
-        "byte-store".to_owned(),
-        options,
-      )?,
-    ))
-  } else {
-    Err(format!(
-      "Cannot initialise remote byte store provider with address {address}, as the scheme is not supported",
-    ))
-  }
-}
 
 #[derive(Clone)]
 pub struct ByteStore {
@@ -74,7 +40,7 @@ impl ByteStore {
 
   pub async fn from_options(options: RemoteOptions) -> Result<ByteStore, String> {
     let instance_name = options.instance_name.clone();
-    let provider = choose_provider(options).await?;
+    let provider = choose_byte_store_provider(options).await?;
     Ok(ByteStore::new(instance_name, provider))
   }
