+++ src/rust/engine/process_execution/remote/Cargo.toml
@@ -37,6 +37,7 @@ strum = { workspace = true }
 strum_macros = { workspace = true }
 parking_lot = { workspace = true }
 opendal = { workspace = true }
+remote_provider = { path = "../../remote_provider" }
 remote_provider_traits = { path = "../../remote_provider/remote_provider_traits" }
 remote_provider_opendal = { path = "../../remote_provider/remote_provider_opendal" }
